+++ src/robopilot/spec/__init__.py
@@ -1,0 +1,13 @@
+"""ProjectSpec serialization and validation helpers."""
+
+from robopilot.spec.io import load_spec, spec_to_yaml, write_spec
+from robopilot.spec.validator import SpecValidationResult, validate_spec
+
+__all__ = [
+    "SpecValidationResult",
+    "load_spec",
+    "spec_to_yaml",
+    "validate_spec",
+    "write_spec",
+]
+